requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
google-generativeai==0.3.2
drf-spectacular==0.27.0